+++ sdltutorials/Camera.h
@@ -17,10 +17,10 @@ class Camera{
 public:
 	static Camera cameraControl; //To access camera from anywhere
 private:
-	int x; //Camera position (in pixel) Note: is actually the upperLeftCorner
-	int y; //Upper Left Corner Y
-	int* targetX; //To target an element (Null if not target)
-	int* targetY;
+	float x; //Camera position (in pixel) Note: is actually the upperLeftCorner
+	float y; //Upper Left Corner Y
+	float* targetX; //To target an element (Null if not target)
+	float* targetY;
 public:
 	int targetMode;
 
@@ -58,7 +58,7 @@ public:
 	*/
 	int getY();
 
-	void setPosition(int posX, int posY);
-	void setTarget(int* posX, int* posY);
+	void setPosition(float posX, float posY);
+	void setTarget(float* posX, float* posY);
 };
